+++ packages/botonic-react/src/util/webchat.ts
@@ -5,7 +5,7 @@ import { v7 as uuidv7 } from 'uuid'
 import { WEBCHAT } from '../constants'
 import { timeZoneToCountryCode } from '../time-zone-to-country-code'
 import { ClientUser } from '../webchat/context/types'
-import { ThemeProps } from '../webchat/theme/types'
+import { WebchatTheme } from '../webchat/theme/types'
 import { getProperty } from './objects'
 
 /**
@@ -14,7 +14,7 @@ import { getProperty } from './objects'
  * If property doesn't exist, returns the defaultValue.
  */
 export const _getThemeProperty =
-  (theme: ThemeProps) => (property: string, defaultValue?: any) => {
+  (theme: WebchatTheme) => (property: string, defaultValue?: any) => {
     for (const [key, value] of Object.entries(WEBCHAT.CUSTOM_PROPERTIES)) {
       if (value === property) {
         const nestedProperty = getProperty(theme, value)