+++ packages/botonic-react/src/webchat/webview/styles.ts
@@ -11,7 +11,17 @@ export const StyledWebview = styled.div`
   height: 100%;
   background-color: ${COLORS.SOLID_WHITE};
   z-index: 2;
-  border-radius: ${WEBCHAT.DEFAULTS.BORDER_RADIUS_TOP_CORNERS};
+  border-radius: ${props => props.theme.style.borderRadius};
+
+  @media (max-width: ${props => props.theme.mobileBreakpoint}px) {
+    position: ${props => props.theme.mobileStyle.position};
+    right: ${props => props.theme.mobileStyle.right};
+    bottom: ${props => props.theme.mobileStyle.bottom};
+    width: ${props => props.theme.mobileStyle.width};
+    height: ${props => props.theme.mobileStyle.height};
+    border-radius: ${props => props.theme.mobileStyle.borderRadius};
+    font-size: ${props => props.theme.mobileStyle.fontSize};
+  }
 `
 
 export const StyledWebviewHeader = styled.div`
@@ -20,7 +30,11 @@ export const StyledWebviewHeader = styled.div`
   background-color: ${COLORS.WILD_SAND_WHITE};
   border-top: 1px solid ${COLORS.SOLID_BLACK_ALPHA_0_2};
   border-bottom: 1px solid ${COLORS.SOLID_BLACK_ALPHA_0_2};
-  border-radius: ${WEBCHAT.DEFAULTS.BORDER_RADIUS_TOP_CORNERS};
+  border-radius: ${props => props.theme.header?.style?.borderRadius};
+
+  @media (max-width: ${props => props.theme.mobileBreakpoint}px) {
+    border-radius: 0px;
+  }
 `
 export const StyledCloseHeader = styled.div`
   display: inline-block;