+++ packages/botonic-react/CHANGELOG.md
@@ -10,14 +10,16 @@ All notable changes to Botonic will be documented in this file.
     Click to see more.
   </summary>
   
-## [0.32.0] - 2025-mm-dd
+## [0.35.0] - 2025-mm-dd
 
 ### Added
 
 ### Changed
 
-- Refactor components used to render a [webview inside the webchat](https://github.com/hubtype/botonic/pull/2976)
-- When channel is not a webchat [Multichannel Carousel](https://github.com/hubtype/botonic/pull/2979) displays an image followed by a text with buttons for each element.
+[PR-3003](https://github.com/hubtype/botonic/pull/3003): ReactBot class with typescript
+[PR-2991](https://github.com/hubtype/botonic/pull/2991): Create Provider WebchatTheme with typescript and add a default theme
+[PR-2995](https://github.com/hubtype/botonic/pull/2995): Add Button style in WebchatTheme
+[PR-2996](https://github.com/hubtype/botonic/pull/2996): ButtonDisabler in typescript
 
 ### Fixed